Add new login form styled
