+++ src/config/config.json
@@ -1,4 +1,4 @@
 {
-  "serverURL": "http://192.168.0.62:4000/",
+  "serverURL": "http://100.120.89.13:4000/",
   "maxLength": 30
 }
