+++ src/css/App.css
@@ -206,9 +206,9 @@ a {
 .loginModal {
     width: 40%;
     background: white;
-    height: 60%;
+    height: 70%;
     margin-left: 30%;
-    margin-top: 8%;
+    margin-top: 3%;
     text-align: center;
     padding: 20px;
     overflow: auto;
@@ -285,6 +285,10 @@ a {
 
 .loginModal input:focus {
     outline: none;
+}
+
+.loginModal a {
+    cursor: pointer!important;
 }
 
 .loginModal input {
@@ -767,6 +771,21 @@ button:focus, button:focus {
     }
 }
 
+@media only screen and (max-width: 1500px) {
+    /* Login */
+    .loginModal {
+        width: 50%;
+        background: white;
+        height: 80%;
+        margin-left: 25%;
+        margin-top: 3%;
+        text-align: center;
+        padding: 20px;
+        overflow: auto;
+        color: #555;
+    }
+}
+
 /* Small devices with little vertical real estate */
 @media only screen and (max-height: 700px) {
     html, body {
@@ -823,6 +842,38 @@ button:focus, button:focus {
     }
     .reviewVerify p {
         padding: 5px;
+    }
+    /* Login */
+    .loginModal {
+        width: 80%;
+        font-size: 8pt;
+        background: white;
+        height: 90%;
+        margin-left: 10%;
+        margin-top: 8%;
+        text-align: center;
+        padding: 10px!important;
+        overflow: auto;
+        color: #555;
+    }
+    .loginModal h3 {
+        font-size: 16px!important;
+    }
+    .google p {
+        margin-top: 2px!important;
+        margin-left: 10px;
+        font-size: 9pt!important;
+    }
+    .facebook p {
+        margin-top: 2px!important;
+        margin-left: 10px;
+        font-size: 9pt!important;
+    }
+    .loginModal img {
+        width: 20px;
+    }
+    .loginModal input {
+        width: 70%!important;
     }
 }
 
@@ -972,7 +1023,6 @@ button:focus, button:focus {
         margin-top: 5%;
     }
     /* Login */
-
     .loginModal {
         width: 80%;
         background: white;
@@ -984,13 +1034,11 @@ button:focus, button:focus {
         overflow: auto;
         color: #555;
     }
-
     .loginModal  h3 {
         font-size: 29px;
         letter-spacing: 2px;
         font-weight: normal;
     }
-
     .google {
         background-color: #fff;
         display: inline-flex;
@@ -1079,7 +1127,6 @@ button:focus, button:focus {
         cursor: pointer;
         outline: none
     }
-
     .signUpText {
         padding: 30px;
         color: #555;
